==> Step 0. Data Preparation.ipynb <==
# Data Preparation
The data source we have been provided with is in JSON format. In this short notebook we'll read the data and save it as a CSV file for simpler loading in later steps.

# required library imports

import pandas as pd

## Reading Data from JSON Object File

# define data file name and relative data path (according the current path of this notebook)
DATA_PATH = './scripts/data'
FILE_NAME = 'News_Category_Dataset_v2.json.zip'

df_dataset = pd.read_json(f'{DATA_PATH}/{FILE_NAME}', orient='records', compression='zip', lines=True)

df_dataset.head()

## Saving Data to File in CSV Format

SAVE_NAME = 'dataset_all.csv.gz'

df_dataset.to_csv(f'{DATA_PATH}/{SAVE_NAME}', compression='gzip', index=False)
