==> Step 2. Data Cleaning & Splitting.ipynb <==
# Data Cleaning & Splitting

During this step we'll be doing some clean-up based on the insights we acquired through __EDA__. We are going to split data into two representative parts (train & test) as well, since a separate test dataset is not provided; eventually we'll save the cleaned data files separately to facilitate a hassle-free loading for the next phase (model training).

# required library imports & initial settings

import re
import string
import wordninja
import pandas as pd
from sklearn.model_selection import train_test_split


RANDOM_SEED = 1024
test_size_fraction = 0.2
column_target = 'category'

## Reading Raw Data

# define relative data path (according the current path of this notebook) and data file name
DATA_PATH = './scripts/data'
FILE_NAME = 'dataset_all.csv.gz'

df_all = pd.read_csv(f'{DATA_PATH}/{FILE_NAME}')
df_all

## Dealing with Missing Values

# remove samples with missing values in 'headline' column
df_all = df_all.dropna(subset=['headline']).reset_index(drop=True)

## Fixing Inconsistencies of Target Column Classes

# create a category tranformer function based on lexical similarity scores and the intuition we have after EDA
def transform_category(value):
    if value in ['ARTS', 'CULTURE & ARTS']:
        return 'ARTS & CULTURE'
    elif value in ['THE WORLDPOST', 'WORLDPOST']:
        return 'WORLD NEWS'
    elif value in ['STYLE']:
        return 'STYLE & BEAUTY'
    elif value in ['PARENTS']:
        return 'PARENTING'
    elif value in ['TASTE']:
        return 'FOOD & DRINK'
    elif value in ['GREEN', 'ENVIRONMENT']:
        return 'GREEN & ENVIRONMENT'
    else:
        return value


df_all[column_target] = df_all.apply(lambda row: transform_category(row[column_target]), axis=1)

## Dropping Unnecessary Columns

columns_desired = ['headline', 'category']

df_all = df_all[columns_desired]
df_all

## Splitting Data into Representative Train & Test Sets
This should take place before going any further; because we will be doing some additional data cleaning which should only be done on the training data.

# let's shuffle the whole dataframe before proceeding
df_all = df_all.sample(frac=1, random_state=RANDOM_SEED).reset_index(drop=True)
# second shuffle with an exponential random seed :D
df_all = df_all.sample(frac=1, random_state=(RANDOM_SEED**2)).reset_index(drop=True)

# split the data into representative training and test sets
df_train_full, df_test = train_test_split(df_all, test_size=test_size_fraction, stratify=df_all[column_target], random_state=RANDOM_SEED)

# reset the index of the dataframes
df_train_full.reset_index(drop=True, inplace=True)
df_train_full.reset_index(drop=True, inplace=True)

__Note:__ You may have heard many times that for final training, we should use the entire dataset. As a side note, we'll be applying all train set considerations to this separate entire set as well; so that we can have an entire dataset ready to use for the final training. No test set exists in this scenario, and the test cases would be unseen out of sample data.

## Preprocessing Train Data
We need to do some preprocessing on the training data, which will help us further clean the training samples.

def preprocess_text(text: str) -> str:
    '''
    Applies text preprocessing to input text.

    Args:
        text (str): example => 'This, is the #TEXT   that needs   to be preprocessed.  '

    Returns:
        str: example => 'this is the text that needs to be preprocessed'
    '''
    text = text.lower()  # convert to lowercase
    text = re.sub('(#)(\S+)', r' \2', text)  # remove hashtags sign
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)  # remove punctuations
    text = re.sub(' +', ' ', text)  # replace multiple whitespaces with a single space
    text = text.strip()  # remove leading and trailing whitespaces
    return text


# create a new column in train dataframe which contains headline in preprocessed form
df_train_full['headline_preprocessed'] = df_train_full.apply(lambda row: preprocess_text(str(row['headline'])), axis=1)
# do the same for our separate 'entire' train dataframe
df_all['headline_preprocessed'] = df_all.apply(lambda row: preprocess_text(str(row['headline'])), axis=1)

# re-order columns
columns_order = ['headline', 'headline_preprocessed', 'category']
df_train_full = df_train_full[columns_order]
df_all = df_all[columns_order]


df_train_full.head()

## Apply Further Data Cleaning on Train Data
This is where we check character length and word count of the headlines and remove any rows that do not meet the requirements we identified during the __EDA__.

def split_concatenated(text: str) -> str:
    '''
    splits words of a concatenated English text string.

    Args:
        text (str): example => 'thistextstringisconcatenated'

    Returns:
        str: example => 'this text string is concatenated'
    '''
    text = ' '.join(wordninja.split(text))
    return text


# apply splitting function to headlines with just 1 word
df_train_full['headline_preprocessed'] = df_train_full['headline_preprocessed'].map(lambda value: split_concatenated(value) if (len(str(value).split()) < 2) else value)
# apply the same process to our separate 'entire' train dataframe
df_all['headline_preprocessed'] = df_all['headline_preprocessed'].map(lambda value: split_concatenated(value) if (len(str(value).split()) < 2) else value)

# drop remaining rows with only 1 word in headline
df_train_full = df_train_full.loc[~df_train_full['headline_preprocessed'].str.split().str.len() < 2].reset_index(drop=True)
# the same for our separate 'entire' train dataframe
df_all = df_all.loc[~df_all['headline_preprocessed'].str.split().str.len() < 2].reset_index(drop=True)

# drop samples having headlines with character length <= 10 and words count above 2
df_train_full = df_train_full.loc[~(df_train_full['headline_preprocessed'].str.len() <= 10) & (df_train_full['headline_preprocessed'].str.split().str.len() > 2)].reset_index(drop=True)
# the same for our separate 'entire' train dataframe
df_all = df_all.loc[~(df_all['headline_preprocessed'].str.len() <= 10) & (df_all['headline_preprocessed'].str.split().str.len() > 2)].reset_index(drop=True)

## Saving Cleaned Versions of Data

df_train_full.to_csv(f'{DATA_PATH}/train_cleaned.csv.gz', compression='gzip', index=False)
df_test.to_csv(f'{DATA_PATH}/test_cleaned.csv.gz', compression='gzip', index=False)
df_all.to_csv(f'{DATA_PATH}/train_cleaned_entire.csv.gz', compression='gzip', index=False)
